% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdftracks-package.R
\docType{package}
\name{mdftracks-package}
\alias{mdftracks}
\alias{mdftracks-package}
\title{mdftracks: Read and Write 'MTrackJ Data Files'}
\description{
'MTrackJ' is an 'ImageJ' plugin for motion tracking and analysis (see \url{https://imagescience.org/meijering/software/mtrackj/}). This package reads and writes 'MTrackJ Data Files' ('.mdf', see \url{https://imagescience.org/meijering/software/mtrackj/format/}). It supports 2D data and read/writes cluster, point, and channel information. If desired, generates track identifiers that are unique over the clusters. See the project page for more information and examples.
}
\seealso{
\href{https://imagescience.org/meijering/software/mtrackj/format/}{MTrackJ Data Format}
}
\author{
\strong{Maintainer}: Gerhard Burger \email{burger.ga@gmail.com} (\href{https://orcid.org/0000-0003-1062-5576}{ORCID})

}
\keyword{internal}
