% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdftracks.example.data}
\alias{mdftracks.example.data}
\title{Example data to show mdftracks functionality}
\format{A data frame with 10 rows and 9 variables
\describe{
\item{cl}{track cluster}
\item{id}{track identifier (in cluster)}
\item{p}{point in track (not necessarily the same as frame number)}
\item{x}{x-coordinate of point}
\item{y}{y-coordinate of point}
\item{z}{z-coordinate of point}
\item{t}{time of point}
\item{ch}{track channel}
\item{uid}{track identifier (unique over clusters)}
}}
\source{
Self-generated
}
\usage{
mdftracks.example.data
}
\description{
Example data to show mdftracks functionality
}
\keyword{datasets}
