% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdftracks.R
\name{write.mdf}
\alias{write.mdf}
\title{Write an MTrackJ track file (.mdf)}
\usage{
write.mdf(x, file = "", cluster.column = NA, id.column = 1,
  time.column = 2, pos.columns = c(3, 4, 5), fileEncoding = "")
}
\description{
Write an MTrackJ track file (.mdf)
}
\examples{
\dontrun{
write.mdf(x, '~/mdftracks.mdf')
}
}
\seealso{
Other mdftracks functions: \code{\link{read.mdf}}
}
