% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdftracks.R
\docType{package}
\name{mdftracks}
\alias{mdftracks}
\alias{mdftracks-package}
\title{mdftracks:  Read MTrackJ Tracks}
\description{
Reads the .mdf files generated by the MTrackJ and convert it to a data.frame.
}

