% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdftracks.R
\docType{package}
\name{mdftracks}
\alias{mdftracks}
\alias{mdftracks-package}
\title{mdftracks:  Read and Write MTrackJ Data Format Files}
\description{
Reads and writes '.mdf' (MTrackJ Data Format) files.
Supports clusters, 2D data, and channel information. If desired, generates
unique track identifiers based on cluster and id data from the '.mdf' file.
}
\seealso{
\href{https://imagescience.org/meijering/software/mtrackj/format/}{MTrackJ Data Format}
}
