% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdftracks.R
\name{read.mdf}
\alias{read.mdf}
\title{Read an MTrackJ track file (.mdf)}
\usage{
read.mdf(file, drop.Z = F, include.point.numbers = FALSE,
  include.channel = F)
}
\arguments{
\item{file}{The MTrackJ .mdf file with the tracks.}
}
\description{
Reads the .mdf files generated by the MTrackJ and convert it to a data.frame.
}
\examples{
\dontrun{
read.mdf('~/mdftracks.mdf')
}
}

